Add service to mount rootfs as tmpfs in initrd if systemd-repart fails
When the install media is read-only, systemd-repart will fail to create
partitions on boot, including the root partition. This will cause
sysroot.mount to fail. An easy workaround is to mount the rootfs as a
tmpfs instead, which allows the installer to run and then reboot into
the normal system environment.

This is a bit hacky; the initrd-tmpfs-root service masks two files
created by systemd-gpt-auto-generator so systemd doesn't try to mount
a non-existent sysroot partition.

The much easier approach of appending "root=tmpfs" to the kernel boot
line isn't feasible because that would also apply to an installed
system, which is obviously not what we want!

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>