Add service to import partitions from CDROM
When running the installer from a CDROM, the kernel won't automatically
detect partitions from the disk. This adds a simple service to run
kpartx to make those partitions available.

The systemd-repart service overrides are extended to no longer count
exit status 77 as success, since we want the failure when there's not a
GPT table present to trigger mounting rootfs as tmpfs.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>