incus-osd/systemd: Set the same recovery/user-supplied passwords on swap as root
When a new secure boot signing key is used for a UKI, we must un-enroll
the existing TPM token on both volumes, then re-enroll with the updated
public key. Because the new public key will invalidate the existing TPM
state, we must rely on the recovery password being set on both the
root and swap LUKS volumes.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>