scripts/test: Add scripts to generate test Secure Boot keys
These scripts generate a full mock-up CA certificate hierarchy similar
to the production certs used when publishing IncusOS. A total of four
Secure Boot signing keys are created, with the first two used to
populate an initial db of trusted certificates. The third is prepared
as an update to db, while the fourth is prepared as an update to dbx.

The resulting keys generated with these scripts are for TEST purposes
only -- don't let them anywhere near a production system.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>