incus-osd/systemd: Check for and add swap recovery key if needed
Installs of IncusOS prior to version 202506220351 won't have a recovery
key set for their swap partition. However, during a Secure Boot key
update logic will expect to be able to use the recovery key to update
the expected PCR7 bindings for the swap volume.

This commit adds logic that checks if the swap volume is missing a
recovery key, and if so, automatically applies the key from the current
state.

This code should be relatively short lived, and removed by the end of
August, 2025.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>