incus-osd/install: Check symlinks to determine install target
A device may have more than one symlink under /dev/disk/by-id/ pointing to it.
If the device ID provided in the install seed is complete, we can now attempt
to resolve that symlink and the symlink of the potential target device. If they
are the same, then return the found target device.

The potential install target device IDs presented to the user originate from
the output of running `lsblk -o ID_LINK` on the system.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>