incus-osd/systemd: Improve wait logic for network device renaming
The prior logic simply checked the journal entries since last boot checking that
at least one physical network device had been renamed as expected. However, if a
system would receive a network update at a later point that referenced a network
device not in the initial configuration, we wouldn't properly wait for udev to
rename it first, causing the network reconfiguration to fail and leave the
system in a bad state.

Now we look at a network configuration and determine which (if any) physical
network devices are new, and then wait for all of them to be renamed in the
recent past. Hopefully this will be more robust than what we had before.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>