base: Use shorter interface names for USB NICs
systemd defaults to naming USB NICs enx<mac>, which is 15 characters long.
However, that's the max length for an interface name, and when we add our prefix
during network config generation, we would end up with a broken state. So,
change the default to use an interface name like enp0s0f6u4.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>