incus-osd/storage: Fix bug in IsRemoteDevice() when given a partition
Adding logic to handle resolving a symlink if provided exposed a bug where we
would fail if the device was a partition and not an entire block device. This
commit adds some simple logic to attempt to detect if the device is a partition,
and if so, then get the actual device it belongs to.

Signed-off-by: Mathias Gibbens <mathias.gibbens@futurfusion.io>